requirements.txt
uvicorn
fastapi
jinja2
deta
python-dotenv
python-multipart
requests
aiofiles
sentry-sdk
boto3